프로젝트 모듈 구조를 정리한다
