--- lib/person.rb
@@ -1,12 +1,0 @@
-class Person
-  attr_reader :schedule
-
-  def initialize name
-    @name = name
-    @schedule = []
-  end
-
-  def add_schedule scd
-    @schedule << scd
-  end
-end
