+++ lib/recurring/set_expr/collection.rb
@@ -1,0 +1,4 @@
+module Recur
+  class Collection
+  end
+end
