+++ lib/recurring/set_expr/intersection.rb
@@ -1,0 +1,22 @@
+require_relative './collection'
+
+module Recur
+  class Intersection < Collection
+    attr_reader :elems
+
+    def initialize
+      @elems = []
+    end
+
+    def add_elem elem
+      @elems << elem
+    end
+
+    def includes date
+      @elems.each do |tmpr_expr|
+        return false if !tmpr_expr.includes(date)
+      end
+      true
+    end
+  end
+end
