+++ lib/recurring/tmpr_expr/temporal_expr.rb
@@ -1,0 +1,7 @@
+module Recur
+  class TemporalExpr
+    def includes date
+      raise 'this method should be overriden'
+    end
+  end
+end
