--- lib/set_expr/collection.rb
@@ -1,2 +1,0 @@
-class Collection
-end
