--- lib/tmpr_expr/temporal_expr.rb
@@ -1,5 +1,0 @@
-class TemporalExpr
-  def includes date
-    raise 'this method should be overriden'
-  end
-end
