+++ spec/recurring/schedule_elem_spec.rb
@@ -1,17 +1,17 @@
-require_relative './helpers/spec_helper'
+require_relative '../helpers/spec_helper'
 
 describe "schedule element" do
   it "각 event에 대한 schedule element를 schedule 객체에 전달한다"
 
   it "주어진 기간 내에 찾으려는 이벤트가 예정된 날짜 목록을 알 수 있다" do
-    scd_elem = ScheduleElem.new(
-      Event.new("1st mon of month", "golf games"),
-      DayInMonth.new(1, 1)
+    scd_elem = Recur::ScheduleElem.new(
+      Recur::Event.new("1st mon of month", "golf games"),
+      Recur::DayInMonth.new(1, 1)
     )
 
     expect(
       scd_elem.occurrences(
-        Event.new("1st mon of month", "golf games"),
+        Recur::Event.new("1st mon of month", "golf games"),
         [Date.new(2024,11,3), Date.new(2024,12,5)]
       )
     ).to eq(
@@ -31,72 +31,72 @@ describe "schedule element" do
   end
 
   it "주어진 날짜 다음에 찾으려는 이벤트가 예정된 날짜를 알 수 있다" do
-    scd_elem = ScheduleElem.new(
-      Event.new("2nd tue of month", "gastro clinic"),
-      DayInMonth.new(2, 2)
+    scd_elem = Recur::ScheduleElem.new(
+      Recur::Event.new("2nd tue of month", "gastro clinic"),
+      Recur::DayInMonth.new(2, 2)
     )
 
     # allow(scd_elem).to receive(:next_occurrence).and_return("2024-11-12")
     expect(
       scd_elem.next_occurrence(
-        Event.new("2nd tue of month", "gastro clinic"),
+        Recur::Event.new("2nd tue of month", "gastro clinic"),
         Date.new(2024,11,10)
       )
     ).to eq(Date.new(2024,11,12))
   end
 
   it "주어진 날짜에 찾으려는 이벤트가 예정되어 있는지 알 수 있다" do
-    scd_elem = ScheduleElem.new(
-      Event.new("1st mon of month", "golf games"),
-      DayInMonth.new(1, 1)
+    scd_elem = Recur::ScheduleElem.new(
+      Recur::Event.new("1st mon of month", "golf games"),
+      Recur::DayInMonth.new(1, 1)
     )
 
     expect(
       scd_elem.is_occurring(
-        Event.new("1st mon of month", "golf games"),
+        Recur::Event.new("1st mon of month", "golf games"),
         Date.new(2024, 12, 2)
       )
     ).to eq(true)
   end
 
   it "tmpr_expr에 set_expr이 올 수 있다" do
-    union = Union.new
-    union.add_elem DayInMonth.new(1, 1)
-    union.add_elem DayInMonth.new(1, 3)
+    union = Recur::Union.new
+    union.add_elem Recur::DayInMonth.new(1, 1)
+    union.add_elem Recur::DayInMonth.new(1, 3)
 
-    scd_elem = ScheduleElem.new(
-      Event.new("1st or 3rd mon of month", "golf games"),
+    scd_elem = Recur::ScheduleElem.new(
+      Recur::Event.new("1st or 3rd mon of month", "golf games"),
       union
     )
 
     [Date.new(2024, 12, 2), Date.new(2024, 12, 16)].each do |d|
       expect(
         scd_elem.is_occurring(
-          Event.new("1st or 3rd mon of month", "golf games"),
+          Recur::Event.new("1st or 3rd mon of month", "golf games"),
           d
         )
       ).to eq(true)
     end
 
 
-    intersection = Intersection.new
-    intersection.add_elem(DayInMonth.new(1, -1))
-    intersection.add_elem(RangeEachYear.new(5, 5))
+    intersection = Recur::Intersection.new
+    intersection.add_elem(Recur::DayInMonth.new(1, -1))
+    intersection.add_elem(Recur::RangeEachYear.new(5, 5))
 
     memorial_days = [[2022, 5, 30], [2023, 5, 29], [2024, 5, 27]]
                            .inject([]) { |res, y_m_d|
                              res << Date.new(y_m_d[0], y_m_d[1], y_m_d[2])
                            }
 
-    scd_elem = ScheduleElem.new(
-      Event.new("last monday in May", "The US holiday of memorial day"),
+    scd_elem = Recur::ScheduleElem.new(
+      Recur::Event.new("last monday in May", "The US holiday of memorial day"),
       intersection
     )
 
     memorial_days.each {|d|
       expect(
         scd_elem.is_occurring(
-          Event.new("last monday in May", "The US holiday of memorial day"),
+          Recur::Event.new("last monday in May", "The US holiday of memorial day"),
           d
         )
       ).to eq(true)
