+++ spec/recurring/schedule_spec.rb
@@ -1,10 +1,10 @@
-require_relative '../lib/schedule'
-require_relative '../lib/event'
-require_relative './helpers/spec_helper'
+require_relative '../../lib/recurring/schedule'
+require_relative '../../lib/recurring/event'
+require_relative '../helpers/spec_helper'
 
 describe "schedule 객체는" do
   it "주어진 기간 내에 찾으려는 이벤트가 예정된 날짜 목록을 알 수 있다" do
-    scd = Schedule.new
+    scd = Recur::Schedule.new
     scd_elem = double()
     scd.add_elem(scd_elem)
 
@@ -12,7 +12,7 @@ describe "schedule 객체는" do
 
     expect(
       scd.occurrences(
-        Event.new("every mon", "golf games"),
+        Recur::Event.new("every mon", "golf games"),
         [Date.new(2024,11,3), Date.new(2024,12,05)]
       )
     ).to eq(
@@ -22,7 +22,7 @@ describe "schedule 객체는" do
   end
 
   it "주어진 날짜 다음에 찾으려는 이벤트가 예정된 날짜를 알 수 있다" do
-    scd = Schedule.new
+    scd = Recur::Schedule.new
     scd_elem = double()
     scd.add_elem(scd_elem)
 
@@ -30,7 +30,7 @@ describe "schedule 객체는" do
 
     expect(
       scd.next_occurrence(
-        Event.new("2nd tue of month", "gastro clinic"),
+        Recur::Event.new("2nd tue of month", "gastro clinic"),
         Date.new(2024,11,10)
       )
     ).to eq(Date.new(2024,11,12))
@@ -38,7 +38,7 @@ describe "schedule 객체는" do
   end
 
   it "주어진 날짜에 찾으려는 이벤트가 예정되어 있는지 알 수 있다" do
-    scd = Schedule.new
+    scd = Recur::Schedule.new
     scd_elem = double()
     scd.add_elem(scd_elem)
 
@@ -46,7 +46,7 @@ describe "schedule 객체는" do
 
     expect(
       scd.is_occurring(
-        Event.new("every mon", "golf games"),
+        Recur::Event.new("every mon", "golf games"),
         Date.new(2024,12,2)
       )
     ).to eq(true)
