+++ spec/recurring/tmpr_expr/temporal_expr_spec.rb
@@ -1,9 +1,9 @@
 require 'date'
 
-require_relative '../lib/tmpr_expr/temporal_expr'
-require_relative '../lib/tmpr_expr/range_each_year'
-require_relative '../lib/tmpr_expr/day_in_month'
-require_relative './helpers/spec_helper'
+require_relative '../../../lib/recurring/tmpr_expr/temporal_expr'
+require_relative '../../../lib/recurring/tmpr_expr/range_each_year'
+require_relative '../../../lib/recurring/tmpr_expr/day_in_month'
+require_relative '../../helpers/spec_helper'
 
 describe "temporal expression" do
   context "temporal expression 구현 방식을 정의한다" do
@@ -16,7 +16,7 @@ describe "temporal expression" do
       invalid_args = [[0, 1], [8, 1]]
 
       invalid_args.each do |arg|
-        expect {DayInMonth.new(*arg)}.to raise_error(RuntimeError)
+        expect {Recur::DayInMonth.new(*arg)}.to raise_error(RuntimeError)
       end
     end
 
@@ -24,18 +24,18 @@ describe "temporal expression" do
       invalid_args = [[1, 0], [1, 6], [1, -6]]
 
       invalid_args.each do |arg|
-        expect {DayInMonth.new(*arg)}.to raise_error(RuntimeError)
+        expect {Recur::DayInMonth.new(*arg)}.to raise_error(RuntimeError)
       end
     end
 
     it "주어진 date의 요일을 알 수 있다" do
-      monday_in_1st_week = DayInMonth.new(1, 1)
+      monday_in_1st_week = Recur::DayInMonth.new(1, 1)
       monday = Date.new(2024, 11, 11)
       expect(monday_in_1st_week.day_matches monday).to eq(true)
     end
 
     it "첫번째 주를 기준으로 주어진 date가 달의 몇번째 주에 속하는지 알 수 있다" do
-      monday_in_2nd_week = DayInMonth.new(1, 2)
+      monday_in_2nd_week = Recur::DayInMonth.new(1, 2)
 
       nov_1st_mon = Date.new(2024, 11, 4)
       expect(monday_in_2nd_week.week_matches nov_1st_mon).to eq(false)
@@ -45,7 +45,7 @@ describe "temporal expression" do
     end
 
     it "마지막 주를 기준으로 주어진 date가 달의 몇번째 주에 속하는지 알 수 있다" do
-      monday_in_last_week = DayInMonth.new(1, -1)
+      monday_in_last_week = Recur::DayInMonth.new(1, -1)
       nov_last_week = Date.new(2024, 11, 29)
       expect(monday_in_last_week.week_matches nov_last_week).to eq(true)
     end
@@ -57,7 +57,7 @@ describe "temporal expression" do
 
     it "주어진 day_num(date는 해당 달에서 day_num번째 날)이 몇번째 주에 속하는지 알 수 있다" do
       ignore1, ignore2 = 1, 1
-      day_in_mon = DayInMonth.new(ignore1, ignore2)
+      day_in_mon = Recur::DayInMonth.new(ignore1, ignore2)
 
       (1..7).each {|day_in_1st_week|
         expect(day_in_mon.week_in_month day_in_1st_week).to eq(1)
@@ -78,7 +78,7 @@ describe "temporal expression" do
 
     it "주어진 날짜에서, 해당 달의 마지막 날까지 남은 일수를 알 수 있다" do
       ignore1, ignore2 = 1, 1
-      day_in_mon = DayInMonth.new(ignore1, ignore2)
+      day_in_mon = Recur::DayInMonth.new(ignore1, ignore2)
       nov_11 = Date.new(2024, 11, 11)
       expect(day_in_mon.days_left_in_month nov_11).to eq(19)
     end
@@ -88,7 +88,7 @@ describe "temporal expression" do
     # 시작과 끝에 해당하는 day, month를 인자로 받는다
 
     it "months_include: 정의한 시작월과 끝월이 주어진 날짜의 달을 포함하는지 알 수 있다" do
-      mar_to_sep = RangeEachYear.new(3, 9, 14, 12)
+      mar_to_sep = Recur::RangeEachYear.new(3, 9, 14, 12)
       out_of_range = [Date.new(2024, 2, 11), Date.new(2024, 11, 11)]
       out_of_range.each {|d|
         expect(mar_to_sep.months_include d).to eq(false)
@@ -100,7 +100,7 @@ describe "temporal expression" do
     end
 
     it "months_include: 주어진 날짜의 달이 시작월이나 끝월인 경우 포함되지 않았다고 판단한다" do
-      mar_to_sep = RangeEachYear.new(3, 9, 14, 12)
+      mar_to_sep = Recur::RangeEachYear.new(3, 9, 14, 12)
       in_boundary = (1..30).inject([]) {|res, day|
         res << Date.new(2024, 3, day)
         res << Date.new(2024, 9, day)
@@ -111,7 +111,7 @@ describe "temporal expression" do
     end
 
     it "stt_month_includes: 정의한 시작 월일이 주어진 날짜를 포함하는지 알 수 있다" do
-      mar_14_to_sep_12 = RangeEachYear.new(3, 9, 14, 12)
+      mar_14_to_sep_12 = Recur::RangeEachYear.new(3, 9, 14, 12)
 
       mon_not_same_with_stt_mon = Date.new(2024, 4, 30)
       expect(mar_14_to_sep_12.stt_month_includes mon_not_same_with_stt_mon).to eq(false)
@@ -126,7 +126,7 @@ describe "temporal expression" do
     end
 
     it "stt_month_includes: 시작일과 끝일이 0인 경우, 주어진 날짜의 시작 월만 같으면 참이다" do
-      mar_to_sep = RangeEachYear.new(3, 9)
+      mar_to_sep = Recur::RangeEachYear.new(3, 9)
 
       mon_same_with_stt_mon = (1..30).inject([]) {|res, day|
         res << Date.new(2024, 3, day)
@@ -147,7 +147,7 @@ describe "temporal expression" do
     end
 
     it "end_month_includes: 정의한 끝 월이 주어진 날짜를 포함하는지 알 수 있다" do
-      mar_14_to_sep_12 = RangeEachYear.new(3, 9, 14, 12)
+      mar_14_to_sep_12 = Recur::RangeEachYear.new(3, 9, 14, 12)
 
       mons_not_same_with_end_mon = [Date.new(2024, 8, 11), Date.new(2024, 10, 30)]
       mons_not_same_with_end_mon.each {|d|
@@ -165,7 +165,7 @@ describe "temporal expression" do
     end
 
     it "includes: 주어진 날짜가 range every year temporal expr이 정의하는 범위 내에 속하는지 알 수 있다" do
-      mar_14_to_sep_12 = RangeEachYear.new(3, 9, 14, 12)
+      mar_14_to_sep_12 = Recur::RangeEachYear.new(3, 9, 14, 12)
 
       before_range, after_range = Date.new(2024, 2, 28), Date.new(2024, 10, 10)
       falsy = [before_range, after_range]
